requirements.txt
albumentations==1.3.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
# av==10.0.0
black==22.3.0
dotty_dict==1.3.1
fire==0.1.3
imagesize==1.4.1
imgaug==0.4.0
lightning-utilities>=0.3.0
mlflow==2.3.1
monai==1.0.1
numpy==1.23.5
parameterized==0.8.1
Pillow==9.1.1
polars==0.16.16
pre-commit==3.1.1
pydantic==1.10.4
pytorch_lightning==2.0.1
PyYAML==6.0
scipy==1.8.1
SimpleITK==2.2.1
tensorboard==2.11.0
torch==1.13.1
torch-snippets==0.496
torchmetrics==0.10.3
torchvision==0.14.1
typing==3.7.4.3
wandb==0.13.4
